backend/chennai_agent/requirements.txt
# Chennai Smart Agent Requirements

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
openpyxl>=3.1.0  # For Excel file support

# LangChain dependencies (should already be installed)
langchain>=0.3.27
langchain-google-genai>=2.1.12
langchain-core>=0.3.0

# Optional: For enhanced features
# selenium>=4.15.0  # Uncomment for dynamic page scraping
# scrapy>=2.11.0  # Uncomment for advanced scraping
